read from files fix

+++ src/Question.java
@@ -2,9 +2,9 @@ import java.util.HashMap;
 
 public class Question {
     private String questionText;
-    private HashMap<String, HashMap<Character, Float>> answerValues;
+    private HashMap<String, HashMap<Character, Double>> answerValues;
 
-    public Question(String questionText, HashMap<String, HashMap<Character, Float>> answerValues) {
+    public Question(String questionText, HashMap<String, HashMap<Character, Double>> answerValues) {
         this.questionText = questionText;
         this.answerValues = answerValues;
     }
